Reorganize schema folder
