+++ ui/e2e/settings/deleteNamespace.spec.ts
@@ -1,7 +1,7 @@
 import { createNamespace, deleteNamespace } from "../utils/namespace";
 import { expect, test } from "@playwright/test";
 
-import { NamespaceListSchemaType } from "~/api/namespaces/schema";
+import { NamespaceListSchemaType } from "~/api/namespaces/schema/namespace";
 import { getNamespaces } from "~/api/namespaces/query/get";
 import { headers } from "e2e/utils/testutils";
